+++ src/main/cpp/SndCtrl/sndctrl.h
@@ -3,6 +3,7 @@
 #include "JniCaller.h"
 #include "Listeners.h"
 #include "FocusListener.h"
+#include "audiopolicyconfigfactory.h"
 
 struct SDeviceNameId {
     co_ptr<WCHAR> name;
@@ -12,11 +13,16 @@ struct SDeviceNameId {
 class SndCtrl : public DeviceListenerCB
 {
 private:
+    static wstring MMDEVAPI_DEVICE_PREFIX; 
+    static wstring MMDEVAPI_RENDER_POSTFIX;
+    static wstring MMDEVAPI_CAPTURE_POSTFIX;
+
     shared_ptr<JniCaller> pJni;
     unique_ptr<FocusListener> pFocusListener;
     CComPtr<IMMDeviceEnumerator> cpEnumerator;
     unordered_map<wstring, unique_ptr<AudioDevice>> devices;
     StoppingHandle<DeviceListener> cpDeviceListener;
+    IAudioPolicyConfigFactory* pPolicyConfigFactory;
 
 public:
     SndCtrl(JNIEnv* env, jobject obj);
@@ -34,6 +40,8 @@ public:
     void UpdateDefaultDevice(wstring id, EDataFlow dataFlow, ERole role);
     void TriggerAv();
 
+    bool SetPersistedDefaultAudioEndpoint(int pid, EDataFlow flow, wstring deviceId);
+    wstring GetPersistedDefaultAudioEndpoint(int pid, EDataFlow flow);
 private:
     void InitDevices();
 
@@ -44,6 +52,8 @@ private:
     SDeviceNameId DeviceNameId(IMMDevice& device);
     CComPtr<IAudioEndpointVolume> GetVolumeControl(IMMDevice& device);
     CComPtr<IAudioSessionManager2> Activate(IMMDevice& device);
+
+    void BuildAudioPolicyConfigFactory();
 };
 
 extern unique_ptr<SndCtrl> pSndCtrl;